pyproject.toml
[project]
name = "hpc-tracker"
version = "0.1.0"
description = "Personal HPC dashboard for job management, monitoring, and productivity"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx",
    "websockets",
    "pyyaml",
    "aiosqlite",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"
